# src/server/swagger/swagger.json
{
  "openapi" : "3.0.2",
  "info" : {
    "title" : "Network Detection and Response (NDR)",
    "description" : "Advanced Cybersecurity Analytics Service (ACAS) for AI-based analysis with explainable intelligence of network events captured by the Montimage Monitoring Tools (MMT)",
    "termsOfService" : "https://github.com/Montimage/acas",
    "contact" : {
      "email" : "developer@montimage.com"
    },
    "license" : {
      "name" : "Montimage",
      "url" : "https://www.montimage.eu"
    },
    "version" : "v1"
  },
...
  "servers": [
    {
      "description": "Localhost Server",
      "url": "http://localhost:31057/api/"
    },
    {
      "description": "Production Server",
      "url": "https://ndr.montimage.eu/api/"
    }
  ],
...
  "paths" : {
...
    "/pcaps/" : {
      "get" : {
        "tags" : [ "pcaps" ],
        "summary" : "Get the list of all pcap files",
        "description" : "Get the list of all pcap files which are available for analyzing",
        "operationId" : "getPcapList",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "pcaps" : {
                      "type" : "array",
                      "items" : {
                        "type" : "string",
                        "example" : "tcp_bbc-out-of-order.pcap"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "pcaps" ],
        "summary" : "Upload a pcap file",
        "description" : "Upload a pcap file to analyse. Authentication is required in production but optional for development/testing.",
        "operationId" : "uploadPcap",
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "pcapFile" : {
                    "type" : "string",
                    "format" : "binary"
                  }
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "pcapFile" : {
                      "type" : "string"
                    }
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Failed to upload pcap files"
          }
        }
      },
      "delete" : {
        "tags" : [ "pcaps" ],
        "summary" : "Delete all pcaps",
        "description" : "Delete all pcaps",
        "operationId" : "deletePcaps",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Successful delete all uploaded pcaps"
          }
        }
      }
    },
...
    "/pcaps/dataset/{datasetName}" : {
      "get" : {
        "tags" : [ "pcaps" ],
        "summary" : "Get the list of all pcap files of a given dataset",
        "description" : "Get the list of all pcap files of a given dataset which can be analyze in dataset mode",
        "operationId" : "getPcapDataset",
        "parameters" : [ {
          "name" : "datasetName",
          "in" : "path",
          "description" : "Name of the dataset",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "ip-dataset01"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "pcaps" : {
                      "type" : "array",
                      "items" : {
                        "type" : "string",
                        "example" : "ip_ping_local_ip_fragmentation.pcap"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "pcaps" ],
        "summary" : "Upload a list of pcap files to a dataset",
        "description" : "Update a list of pcap files to build a dataset",
        "operationId" : "uploadPcapDataset",
        "parameters" : [ {
          "name" : "datasetName",
          "in" : "path",
          "description" : "Name of the dataset",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "my_dataset"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "pcapFiles" : {
                    "type" : "array",
                    "items" : {
                      "type" : "string",
                      "format" : "binary"
                    }
                  }
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "pcapFileName" : {
                      "type" : "string",
                      "example" : "my_dataset.pcap"
                    }
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Failed to create dataset location"
          }
        }
      },
      "delete" : {
        "tags" : [ "pcaps" ],
        "summary" : "Delete the dataset and its pcap files",
        "description" : "Delete the dataset and its pcap files",
        "operationId" : "deleteDatasetPcaps",
        "parameters" : [ {
          "name" : "datasetName",
          "in" : "path",
          "description" : "Name of the dataset",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "my_dataset"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful delete the dataset and its pcap files"
          }
        }
      }
    },
...
  },
...
}